% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterItems.R
\name{clusterItems}
\alias{clusterItems}
\title{Item cluster approach to partial MI}
\usage{
res_clusterItems <- clusterItems(res_testMI = NULL,
                                        clusterWhat = NULL,
                                        method = NULL,
                                        alphaValue = NULL,
                                        loadThreshold = NULL,
                                        intThreshold = NULL,
                                        ...)
}
\arguments{
\item{res_testMI}{Result object of \code{\link{testMI}}. Alternatively, a MI model
can be described using the arguments of \code{\link{testMI}} (see ...).}

\item{clusterWhat}{String, either \code{"loadings"}, \code{"difficulties"}, or
\code{c("loadings", "difficulties")}. Indicates, for which parameter type MI does not hold.
Usually, this will be indicated per factor in a single string
(e.g. \code{"Factor1 configural Factor2 weak"})}

\item{method}{Criteria used for clustering.
One can either specify the maximum difference in item parameters between any
two items within a cluster \code{"threshold"} or base clustering
on nonsignificance of item parameter differences of any two items within
a cluster \code{"sigTest"}. The threshold approach is based upon the approach described
in Pohl, Schulze, & Stets (in press); Pohl & Schulze (2020) and Schulze & Pohl (2021).
The significance test is used as described in Bechger & Maris (2015).}

\item{alphaValue}{Type I error used for clustering via significance testing.
This argument only needs to be specified if method is \code{"sigtest"}.}

\item{loadThreshold}{numeric value for the loading threshold. Only used if method is
\code{"threshold"} and if applicable in the model.}

\item{intThreshold}{numeric value for the intercept/difficulty threshold. Only used if
method is \code{"threshold"}.}

\item{...}{Arguments of \code{\link{testMI}}, if \code{\link{testMI}} has not
been called before and a MI model is to be described.}
}
\value{
A list containing
\itemize{
 \item Data (filtered for potential missings in the covariate)
 \item Model specification(s)
 \item Clustering results organized per factor. If inherited from
 \code{\link{testMI}}, MI tests are also present.
 }
Usually accessed via summary().
}
\description{
Find subsets of items (clusters) which function homogeneously
and may be possible anchor candidates.  Can currently deal with two-group models
for metric items or dichotomous items (Rasch, 2PL, or probit model).
}
\references{
Bechger, T. M., & Maris, G. (2015). A statistical test for diﬀerential item pair functioning.
Psychometrika, 80(2), 317-340. \cr \cr
Pohl, S., Schulze, D., & Stets, E. (in press). Partial measurement invariance: Extending and
evaluating the cluster approach for identifying anchor items. Applied Psychological Assessment. \cr \cr
Pohl, S., & Schulze, D. (2020). Assessing group comparisons or change over
time under measurement non-invariance: The cluster approach for nonuniform DIF.
Psychological Test and Assessment Modeling, 62(2), 281-303. \cr \cr
Schulze, D., & Pohl, S. (2021). Finding clusters of measurement invariant items for
continuous covariates. Structural Equation Modeling: A Multidisciplinary Journal, 28(2), 219-228.
}
