% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{getModel}
\usage{
res <- getModel(res = NULL,
                       which = NULL,
                       silent = FALSE)
}
\arguments{
\item{res}{Object generated by \code{\link{testMI}} or \code{\link{modelAveraging}}.}

\item{which}{Either
 \itemize{
 \item{String, for MI level in \code{\link{testMI}}. Can be \code{configural}
, \code{weak}, \code{strong}, or \code{strict}.}
 \item{Numeric, cluster number for desired partial MI model in
 \code{\link{modelAveraging}}.}
}}

\item{silent}{Do not print summary output? Defaults to \code{FALSE}.}
}
\value{
Parameters (and model fit for non-Bayesian) of a single unidimensional
model of either a specific MI level or a specific partial MI model.
Results are directly printed to the console but can
be further inspected by functions of lavaan/mirt or blavaan/rstan.
}
\description{
Extract a specific model from global MI testing or from model avaraging.
}
