% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{modelAveraging}
\alias{modelAveraging}
\title{modelAveraging}
\usage{
bma <- modelAveraging(res_clusterItems,
                             weights = c(0.5, 0.5))
}
\arguments{
\item{res_clusterItems}{Object generated by \code{\link{clusterItems}}.}

\item{res_modelAveraging}{Object generated by previous use \code{\link{modelAveraging}}.
Different weights can be applied in a timely manner.}

\item{weights}{Weights in numerical order of the clusters. Have to sum to one.}

\item{cores}{Number of iterations (total: warm-up and sampling).}

\item{silent}{Do not print summary output? Defaults to \code{FALSE}.}
}
\value{
Bayesian Average of the mean difference of two groups ("muAver"). Info on
convergence and precision. All fitted STAN objects.
}
\description{
Apply Bayesian model averaging for aggregating the results of
multiple partial MI models. Uses STAN.
}
\details{
If multiple cores are used (like in the default), STAN will maximize
the Viewer Tab of RStudio.
}
