% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialMI.R
\name{partialMI}
\alias{partialMI}
\title{partial MI model}
\usage{
res <- partialMI(res_clusterItems = NULL,
                        anchor = NULL,
                        partialMIwhat = NULL,
                        silent = FALSE,
                        ...)
}
\arguments{
\item{res_clusterItems}{Optional object generated by \code{\link{clusterItems}}.
Alternatively, a MI model
can be described using the arguments of \code{\link{testMI}} (see ...).}

\item{anchor}{Either a number indicating which cluster from the results of \code{\link{clusterItems}}
to use for anchoring, or a vector of item names which shall serve as anchors.}

\item{partialMIwhat}{String, either \code{"loadings"}, \code{"difficulties"}, or
\code{c("loadings", "difficulties")}. Has not to be specified when using res_clusterItems but when setting up
a new MI model. This is an equivalent to clusterWhat in \code{\link{clusterMI}}.}

\item{silent}{Do not print summary output? Defaults to \code{FALSE}.}

\item{...}{Arguments of \code{\link{testMI}}, if \code{\link{testMI}} or
\code{\link{clusterItems}} has not
been called before to describe a measurement model.}
}
\value{
Parameter estimates of the model as defined by the model argument.
Results are directly printed to the console but can be further inspected by l
avaan's or mirt's functions.
}
\description{
Estimate a single unidimensional partial MI model with a given anchor item set (aka cluster).
Can currently deal with two-group models
for metric items or dichotomous items (Rasch, 2PL, or probit model).
}
