% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAverage.R
\name{plotAverage}
\alias{plotAverage}
\title{plotAverage}
\usage{
plotAverage(res_modelAveraging = NULL)
}
\arguments{
\item{res_modelAveraging}{Object generated by previous use
\code{\link{modelAveraging}}.}

\item{...}{further arguments passed to \code{hist()}, except \code{ylim}.}
}
\value{
A histogram with 95% credibility interval lines.
}
\description{
Plotting the distribution of a parameter resulting from Bayesian
model averaging
}
