% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCluster.R
\name{plotCluster}
\alias{plotCluster}
\title{Plotting item clusters}
\usage{
plotCluster(res = res_clusterItems,
                   showLegend = FALSE,
                   tickWidth = 0.2)
}
\arguments{
\item{res}{Results generated by \code{\link{clusterItems}} or \code{\link{testMI}}.}

\item{showLegend}{Adds item identifiers to the plot. Defaults to \code{FALSE}.}

\item{tickWidth}{Give widths of ticks manually. Defaults to \code{0.2}.}
}
\value{
A plot, either a single plot if MI analysis was done for only one parameter,
or multiple plots showing the subsequent clustering via item loadings and difficulties.
Relative DIF effects are always moved to a scale where the lowest DIF values is zero.
}
\description{
Plotting relative DIFs and item clusters.
}
