% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDIF.R
\name{printDIF}
\alias{printDIF}
\title{Print relative DIF matrices}
\usage{
printDIF(res_clusterItems)
}
\arguments{
\item{res}{Results generated by \code{\link{clusterItems}} or \code{\link{testMI}}.}
}
\value{
Relative DIF matrices for item loadings and difficulties.
}
\description{
Printing relative DIF matrices for a single model or from multiple clustering steps.
}
