% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterItems.R
\name{summary.clusterItems}
\alias{summary.clusterItems}
\title{Summary function for clusterItems}
\usage{
\method{summary}{clusterItems}(res, order = "clusters")
}
\arguments{
\item{order}{Choose wether to order cluster solution by \code{"clusters"} (default)
or \code{"items"}.}
}
\description{
Giving a summary of the found item clusters per factor.
}
