% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMI.R
\name{testMI}
\alias{testMI}
\title{Global MI test}
\usage{
res_testMI <- testMI(items = NULL,
                     data = NULL,
                     group = NULL,
                     MIlevel = NULL,
                     stdItems = TRUE,
                     itemType = "metric",
                     dichModel = "factor")
}
\arguments{
\item{items}{Describes the measurement model. It can either be:
\itemize{
 \item A vector of item names, e.g. \code{c("item1", "item2", "item3", "item4")}
 \item A data frame section containing the items e.g. \code{data[1:4]}
 }
This creates a single factor model out of the item names.}

\item{data}{A data frame containing all the used variables. Missing values have
to be declared as \code{NA}.}

\item{group}{String, indicating the grouping variable.}

\item{MIlevel}{String, indicating the level of MI to be tested. Can be either \code{configural}, \code{weak},
\code{strong}, or \code{strict}.}

\item{stdItems}{Should continuous items be standardized? Defaults to TRUE.
Useful for making DIF sizes comparable.}

\item{itemType}{Can be \code{"dichotomous"} or \code{"metric"} (default).}

\item{dichModel}{String indicating which measurement model is applied in the case of dichotomous items.
Values are \code{"factor"} (default) for probit link via WLSMV in lavaan.
\code{"Rasch"} or  \code{"2PL"} for the respective IRT models to be estimated
with the EM estimator in mirt.}
}
\value{
A list containing
\itemize{
 \item Data (filtered for potential missings in the covariate)
 \item Model specification(s)
 \item Results of the MI tests
 }
Usually accessed via summary() or getModel().
}
\description{
Global measurement invariance (MI) tests for the common MI levels
(configural, weak, strong, and strict).  Can currently deal with multiple-group models (two or more groups)
for metric items or dichotomous items (Rasch, 2PL, or probit model).
}
