% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterItems.R
\name{clusterItems}
\alias{clusterItems}
\title{clusterItems}
\usage{
res_clusterMI <- clusterMI(res_testMI,
                                  MIholding = "Factor1 configural Factor2 weak",
                                  method = "threshold",
                                  loadThreshold = 0.3,
                                  intThreshold = 0.5)
}
\arguments{
\item{res_testMI}{Object generated by \code{\link{testMI}}.}

\item{MIholding}{String, either \code{configural}, \code{weak}, \code{strong}
, or \code{strict} indicating the MI level that holds given the results from
\code{\link{testMI}}.
Usually, this will be indicated per factor in a single string
(e.g. \code{"Factor1 configural Factor2 weak"})}

\item{method}{Clustering approach: Via \code{"threshold"} or significance test
(\code{"sigTest"})?}

\item{pValue}{p value to be used if method is \code{"sigtest"}.}

\item{loadThreshold}{numeric value for the loading threshold if method is
\code{"threshold"}.}

\item{intThreshold}{numeric value for the intercept/difficulty threshold if
method is \code{"threshold"}.}
}
\value{
A list expanding the result from \code{testMI()} with the clustering results.
}
\description{
Find subsets of items (clusters) as anchor candidates for which MI holds.
}
