% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{getModel}
\usage{
res <- getModel(res_testMI,
                       "Factor1",
                       "configural")
}
\arguments{
\item{res_testMI}{Object generated by \code{\link{testMI}}.}

\item{factor}{String, factor name.}

\item{which}{String, which MI level? Either \code{configural}
, \code{weak}
, \code{strong}
, or \code{strict}.}

\item{silent}{Do not print output? Defaults to \code{FALSE}.}
}
\value{
A single (multidimensional) model as defined by the \code{model}
argument in \code{testMI}. Results are directly printed to the console but can
be further inspected by lavaan's or mirt's functions.
}
\description{
Estimate a single partial MI model with a given anchor item set (aka cluster).
}
