% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAveraging.R
\name{modelAveraging}
\alias{modelAveraging}
\title{modelAveraging}
\usage{
bma <- modelAveraging(res_clusterItems = NULL,
                             res_modelAveraging = NULL,
                             weights = NULL,
                             CPUcores = 2,
                             chains = 2,
                             iter = NULL,
                             burninPerc = 0.5,
                             silent = FALSE,
                             partialMIwhat,
                             anchors = NULL,
                             ...)
}
\arguments{
\item{res_clusterItems}{Object generated by \code{\link{clusterItems}}. Delivers the
clustering structure as well as the model setup.}

\item{res_modelAveraging}{Object generated by previous use \code{\link{modelAveraging}}.
An alternative to res_clusterItems.
Different weights can thereby be applied in a timely manner.}

\item{weights}{Weights in numerical order of the clusters. This represents the believe of the researcher in
the plausibility of the item cluster as appropriate anchors (see Schulze & Pohl, 2021).
With no prior knowledge or beliefs, weights should be chosen to be equally distributed across all clusters.
Weights have to sum to one.}

\item{CPUcores}{Number of CPU cores to be used. Defaults to 2.}

\item{chains}{Number of chains. Defaults to 2.}

\item{iter}{Number of total iterations. Defaults to 2'000 for metric items (blavaan)
and 10'000 for dichotomous items (IRT models)}

\item{burninPerc}{Percentage of burn-in iteration of the total iterations. Defaults to 0.5.}

\item{silent}{Do not print summary output? Defaults to \code{FALSE}.}

\item{partialMIwhat}{String, either \code{"loadings"}, \code{"difficulties"}, or
\code{c("loadings", "difficulties")}. Has not to be specified when using res_clusterItems but when setting up
a new MI model. This is an equivalent to clusterWhat in \code{\link{clusterMI}}.
 @param ... Arguments of \code{\link{testMI}}, if \code{\link{testMI}} or
\code{\link{clusterItems}} has not
been called before to describe a measurement model.}

\item{anchors}{A list of vectors of item names which shall serve as anchors.}
}
\value{
Bayesian Average of the mean difference of two groups ("muAver"). Info on
convergence and precision. All fitted STAN objects.
}
\description{
Apply Bayesian model averaging for aggregating the results of
multiple partial MI models. Uses STAN. Can currently deal with two-group models
for metric items or dichotomous items (Rasch or 2PL model).
}
\details{
If multiple cores are used (like in the default), STAN will maximize
the Viewer Tab of RStudio.
}
\references{
Schulze, D., & Pohl, S. (2021). Measurement Invariance: Dealing with the uncertainty
in anchor item choice by model averaging. [Manuscript submitted]
}
