% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialMI.R
\name{partialMI}
\alias{partialMI}
\title{partialMI}
\usage{
res <- partialMI(res_clusterItems,
                        "Factor1 1 Factor2 1")
}
\arguments{
\item{res_clusterItems}{Object generated by \code{\link{clusterItems}}.}

\item{anchorClusters}{# String indicating which cluster to use for which factor,
e.g. \code{"Factor1 1 Factor2 1"}
for indicating to use cluster 1 of both item clusterings}

\item{silent}{Do not print summary output? Defaults to \code{FALSE}.}
}
\value{
A single (multidimensional) model as defined by the \code{model}
argument in \code{testMI}. Results are directly printed to the console but can
be further inspected by lavaan's or mirt's functions.
}
\description{
Estimate a single partial MI model with a given anchor item set (aka cluster).
}
